Support replies in repl.

+++ orderbook/src/commands.rs
@@ -4,6 +4,6 @@ use crate::data::orders;
 #[derive(Debug, Clone, Copy, Deserialize, Serialize)]
 pub enum Cmd {
     Order(orders::Order),
-    Snapshot(i32)
+    Snapshot(usize)
 }
